+++ lib/metro/models/ny_template.dart
@@ -4,10 +4,12 @@ class NyTemplate {
   String saveTo;
   List<String> pluginsRequired = [];
   String stub;
+  Map<String, dynamic> options = {};
 
   NyTemplate(
       {required this.name,
       required this.saveTo,
       required this.pluginsRequired,
-      required this.stub});
+      required this.stub,
+      this.options = const {}});
 }